refactor(dictation): 抽出 dictation_end 子模块（end_session 并发敏感整体搬，不动锁/await）
